basic refactoring clean-up
