+++ reference-mod/src/main/java/dev/mineblock11/fabric/referencemod/MyMod.java
@@ -1,10 +1,17 @@
 package dev.mineblock11.fabric.referencemod;
 
 import com.mojang.blaze3d.systems.RenderSystem;
+import dev.mineblock11.fabric.referencemod.block.ModBlocks;
+import dev.mineblock11.fabric.referencemod.block.entity.ModBlockEntities;
+import dev.mineblock11.fabric.referencemod.event.ModEvents;
+import dev.mineblock11.fabric.referencemod.item.ModItems;
+import dev.mineblock11.fabric.referencemod.particle.ModParticles;
+import dev.mineblock11.fabric.referencemod.sound.ModSounds;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.client.rendering.v1.HudRenderCallback;
 import net.fabricmc.fabric.api.client.rendering.v1.WorldRenderEvents;
 import net.fabricmc.fabric.api.event.player.AttackEntityCallback;
+import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.hud.ChatHud;
 import net.minecraft.client.render.*;
@@ -19,9 +26,12 @@ import net.minecraft.util.math.Vec3d;
 import org.apache.commons.lang3.text.WordUtils;
 import org.joml.Matrix4f;
 import org.lwjgl.opengl.GL11;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class MyMod implements ModInitializer {
     public static final String MOD_ID = "referencemod";
+    public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
 
     public static String generateHumanReadable(Identifier identifier) {
         String identifier_path = identifier.getPath();
@@ -30,20 +40,19 @@ public class MyMod implements ModInitializer {
         return capitalized;
     }
 
+    public static void devLogger(String input) {
+        if (!FabricLoader.getInstance().isDevelopmentEnvironment()) return;
+        LOGGER.info("DEV - [" + input + "]");
+    }
+
     @Override
     public void onInitialize() {
         ModBlocks.initialize();
         ModBlockEntities.registerBlockEntityTypes();
         ModItems.initialize();
         ModSounds.initializeSounds();
+        ModEvents.register();
         ModParticles.initialize();
-
-        PlayerDiedCallback.EVENT.register((player, deathMessage) -> {
-            MinecraftClient client = MinecraftClient.getInstance();
-            ChatHud chat = client.inGameHud.getChatHud();
-
-            chat.addMessage(Text.literal(player.getEntityName() + "... you're an absolute idiot."));
-        });
 
         HudRenderCallback.EVENT.register((drawContext, tickDelta) -> {
             Matrix4f positionMatrix = drawContext.getMatrices().peek().getPositionMatrix();
