+++ reference-mod/src/main/java/dev/mineblock11/fabric/referencemod/block/custom/PrismarineLampBlock.java
@@ -1,4 +1,4 @@
-package dev.mineblock11.fabric.referencemod.block;
+package dev.mineblock11.fabric.referencemod.block.custom;
 
 import net.minecraft.block.AbstractBlock;
 import net.minecraft.block.Block;
