+++ reference-mod/src/main/java/dev/mineblock11/fabric/referencemod/datagen/MyBlockLootTableProvider.java
@@ -1,7 +1,7 @@
 package dev.mineblock11.fabric.referencemod.datagen;
 
-import dev.mineblock11.fabric.referencemod.ModBlocks;
-import dev.mineblock11.fabric.referencemod.ModItems;
+import dev.mineblock11.fabric.referencemod.block.ModBlocks;
+import dev.mineblock11.fabric.referencemod.item.ModItems;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricBlockLootTableProvider;
 import net.minecraft.enchantment.Enchantments;
